Problem: On MacOsX, El-Capitan, the instructions on autopilotpattern/hello-world resulted in a page that displayed the following
404 Not Found
nginx/1.10.1

404 Not Found
nginx/1.10.1

Solution:
Change the docker-compose.yml to version 3 of the file format.

With this change, the output is

Hello World

which is the expected output.

Why does this work?:
This explanation is based on empirical evidence rather than any deep knowledge of the inner workings of docker-compose.

version "3" of the docker-compose configuration seems to handle networks differently. As a consequence this seems to fix the dns issue.